requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-jose[cryptography]
passlib[bcrypt]
sqlalchemy[asyncio]
asyncpg
requests
python-dotenv
psycopg2-binary
cryptography